Combine quest commands
